+++ lib/iron_trail/change_model_concern.rb
@@ -21,6 +21,24 @@ module IronTrail
         _where_object_changes(0, args)
       end
 
+      # Allows filtering out updates that changed just a certain set of columns.
+      # This could be useful, for instance, to filter out updates made with
+      # ActiveRecord's #touch method, which changes only the updated_at column.
+      # In that case, calling `.with_delta_other_than(:updated_at)` would exclude
+      # such changes from the result.
+      #
+      # This works by inspecting whether there are any keys in the rec_delta column
+      # other than the columns specified in the `columns` parameter.
+      #
+      # Caveats: this implicitly filters out insert and delete operations, because
+      #   rec_delta would be null in such cases.
+      def with_delta_other_than(*columns)
+        quoted_columns = columns.map { |col_name| connection.quote(col_name) }
+        exclude_array = "ARRAY[#{quoted_columns.join(', ')}]::text[]"
+
+        where(::Arel::Nodes::SqlLiteral.new("(rec_delta - #{exclude_array}) <> '{}'::jsonb"))
+      end
+
       private
 
       def _where_object_changes(ary_index, args)