+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add '.with_delta_other_than' to `IronTrail::ChangeModelConcern`
+
 ## 0.1.4 - 2025-02-24
 
 ### Fixed